[API] Implement Signin (#5)
## Summary
The user provides a user id and password to signin. The server rehashes
the password and compares it with the user's stored password.

## BatchWriteReq Script
I wrote a script to help load Local DynamoDB with entries needed in
integration tests. Not many entries yet, but should be more helpful in
the future.

## Testing
I tagged integration tests with `integration` since they won't pass on
my local machine. Now the `make test` command excludes integration tests
and the github workflow includes _all_ tests.