[API] Implement GetTask (#7)
## Summary
GetTask uses a task ID to retrieve a task from DDB. It only returns it
to the client if the task's user id matches the requestor's user id.

I don't foresee a lot of scenarios where the user will have a task ID,
or even scenarios where they want to grab just one task. So I'll either
leave this RPC as is, for those unique scenarios, or I'll update it to
respond to more likely scenarios. Example: user wants to grab all tasks
that are due today; user wants to grab all tasks that are tag with
"home".

If I don't update this RPC, I'll just create a new one that performs as
described. BatchGetTask or QueryTasks probably.

## Testing
I realized that test files tagged with `integration` have not been
tested in the github workflow that last two PRs. So I'll have to fix
those tests in this PR. I didn't even write an integration test for
`GetTask` because I don't want this PR to get too bloated.